:sparkles: Implement validate rol

+++ app/Http/Controllers/CategoryController.php
@@ -11,7 +11,7 @@ use Firebase\JWT\Key;
 class CategoryController extends Controller
 {
 
-    public function allCategory()
+    public function allCategory(Request $request)
     { 
         
         $jwt = substr($request->header('Authorization', 'token <token>'), 7);
@@ -20,15 +20,30 @@ class CategoryController extends Controller
             
             JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
 
+            $payload = JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
+            
+            $payloadA = (array) $payload;
+
+            if($payloadA[0]->rol_id == 1 || $payloadA[0]->rol_id == 2 || $payloadA[0]->rol_id == 3) {
+
             $category = Category::all()->toArray();
 
-            return response()->json(
-                [
-                'code' => 200,
-                'status' => 'ok',
-                'data' =>$category
-                ]
+                return response()->json(
+                    [
+                    'code' => 200,
+                    'status' => 'ok',
+                    'data' =>$category
+                    ]
+                    );
+
+            }else{
+                return response()->json(
+                    [
+                    'code'=> 401,
+                    'message' => 'usuario no autorizado'
+                    ]
                 );
+            }    
 
         } catch (\Exception $th) {
 
@@ -44,7 +59,7 @@ class CategoryController extends Controller
         }
     }
 
-    public function findCategoryId($id)
+    public function findCategoryId(Request $request, $id)
     { 
 
         $jwt = substr($request->header('Authorization', 'token <token>'), 7);
@@ -53,15 +68,29 @@ class CategoryController extends Controller
 
             JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
             
+            $payload = JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
+            
+            $payloadA = (array) $payload;
+
+            if($payloadA[0]->rol_id == 1 || $payloadA[0]->rol_id == 2 || $payloadA[0]->rol_id == 3) {
+
             $category = Category::find($id);
 
-            return response()->json(
-                [
-                'code' => 200,
-                'status' => 'ok',
-                'data' =>$category
-                ]
+                return response()->json(
+                    [
+                    'code' => 200,
+                    'status' => 'ok',
+                    'data' =>$category
+                    ]
+                    );
+            }else{
+                return response()->json(
+                    [
+                    'code'=> 401,
+                    'message' => 'usuario no autorizado'
+                    ]
                 );
+            }      
 
         } catch (\Exception $th) {
 
@@ -86,15 +115,30 @@ class CategoryController extends Controller
             
             JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
 
+            $payload = JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
+            
+            $payloadA = (array) $payload;
+
+            if($payloadA[0]->rol_id == 1) {
+
             Category::create($request->all());
 
-            return response()->json(
-                [
-                'code' => 201,
-                'status' => 'ok',
-                'data' => $request->all(),
-                ]
+                return response()->json(
+                    [
+                    'code' => 201,
+                    'status' => 'ok',
+                    'data' => $request->all(),
+                    ]
+                    );
+
+            }else{
+                return response()->json(
+                    [
+                    'code'=> 401,
+                    'message' => 'usuario no autorizado'
+                    ]
                 );
+            } 
 
         } catch (\Exception $th) {
 
@@ -119,19 +163,34 @@ class CategoryController extends Controller
 
             JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
             
+            $payload = JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
+            
+            $payloadA = (array) $payload;
+
+            if($payloadA[0]->rol_id == 1) {
+
             $category = Category::find($id);
 
             $category->description = $request->description;
        
             $category->save();
 
-            return response()->json(
-                [
-                'code' => 201,
-                'status' => 'ok',
-                'data' => $category,
-                ]
+                return response()->json(
+                    [
+                    'code' => 201,
+                    'status' => 'ok',
+                    'data' => $category,
+                    ]
+                    );
+
+            }else{
+                return response()->json(
+                    [
+                    'code'=> 401,
+                    'message' => 'usuario no autorizado'
+                    ]
                 );
+            } 
 
         } catch (\Exception $th) {
 
@@ -147,7 +206,7 @@ class CategoryController extends Controller
         }
     }
 
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
 
         $jwt = substr($request->header('Authorization', 'token <token>'), 7);
@@ -156,16 +215,31 @@ class CategoryController extends Controller
 
             JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
          
+            $payload = JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
+            
+            $payloadA = (array) $payload;
+
+            if($payloadA[0]->rol_id == 1) {
+
             $category = Category::find($id);
         
             $category->delete();
             
-            return response()->json(
-                [
-                'code' => 204,
-                'status' => 'success'
-                ]
+                return response()->json(
+                    [
+                    'code' => 204,
+                    'status' => 'success'
+                    ]
+                    );
+
+            }else{
+                return response()->json(
+                    [
+                    'code'=> 401,
+                    'message' => 'usuario no autorizado'
+                    ]
                 );
+            } 
 
         } catch (\Exception $th) {
