+++ app/Http/Controllers/UserController.php
@@ -100,33 +100,58 @@ class UserController extends Controller
         try {
 
             JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
+           
+            $payload = JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
+            
+            $payloadA = (array) $payload;
 
+            if($payloadA[0]->rol_id == 2 || $payloadA[0]->rol_id == 3) {
 
-            $pass = $request->passsword;
+                $pass = $request->password;
 
-            if($pass == null){
+                if($payloadA[0]->id == $id) {
 
-                $user = User::find($id);
+                    if($pass == null){
 
-                $user->update($request->all());
+                        $user = User::find($id);
+        
+                        $user->update($request->all());
+                    }else{
+        
+                        $user = User::find($id);
+        
+                        $request->request->add([
+                            'password' => Hash::make($request->input('password'))
+                        ]);
+        
+                        $user->update($request->all());
+                    }
+        
+                    return response()->json(
+                        [
+                        'code' => 201,
+                        'status' => 'ok',
+                        'data' => $user,
+                        ]
+                        );
+        
+                }else{
+                    return response()->json(
+                        [
+                        'code'=> 401,
+                        'message' => 'usuario no autorizado'
+                        ]
+                    );
+                    }
+
             }else{
-
-                $user = User::find($id);
-
-                $request->request->add([
-                    'password' => Hash::make($request->input('password'))
-                ]);
-
-                $user->update($request->all());
-            }
-
-            return response()->json(
-                [
-                'code' => 201,
-                'status' => 'ok',
-                'data' => $user,
-                ]
+                return response()->json(
+                    [
+                    'code'=> 401,
+                    'message' => 'usuario no autorizado'
+                    ]
                 );
+            }            
 
         } catch (\Exception $th) {
 
@@ -142,7 +167,7 @@ class UserController extends Controller
         }
     }
 
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
 
         $jwt = substr($request->header('Authorization', 'token <token>'), 7);
@@ -151,16 +176,31 @@ class UserController extends Controller
          
             JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
 
+            $payload = JWT::decode($jwt, new Key(env('JWT_SECRET'), 'HS256'));
+            
+            $payloadA = (array) $payload;
+
+            if($payloadA[0]->rol_id == 1) {
+
             $user = User::find($id);
         
             $user->delete();
             
-            return response()->json(
-                [
-                'code' => 204,
-                'status' => 'success'
-                ]
+                return response()->json(
+                    [
+                    'code' => 204,
+                    'status' => 'success'
+                    ]
+                    );
+
+            }else{
+                return response()->json(
+                    [
+                    'code'=> 401,
+                    'message' => 'usuario no autorizado'
+                    ]
                 );
+            } 
 
         } catch (\Exception $th) {
